requirements.txt
asyncio~=3.4.3
aiogram~=3.15.0
aiohttp~=3.11.10
certifi~=2024.8.30
torch~=2.2.0
torchvision~=0.17.0
Pillow~=10.0.0